requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
charset-normalizer==2.0.1
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.2
Django==3.2.5
django-debug-toolbar==3.2.1
django-dotenv==1.4.2
django-rest-framework==0.1.0
djangorestframework==3.12.4
drf-yasg==1.20.0
drf-spectacular==0.23.1
filelock==3.0.12
idna==3.2
importlib-metadata==4.6.1
importlib-resources==5.2.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==21.0
platformdirs==2.0.2
psycopg2==2.9.1
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
virtualenv==20.5.0
zipp==3.5.0
python-dotenv==0.20.0
django-oauth-toolkit==2.1.0
